# oftbackend/openapi.yaml
openapi: 3.1.0
info:
  title: oftbackend API
  version: 1.0.0
servers:
  - url: https://api.oftbackend.example
paths:
  /healthz:
    get:
      summary: Liveness probe
      responses:
        "200":
          description: Healthy
  /readyz:
    get:
      summary: Readiness probe
      responses:
        "200":
          description: Ready
        "503":
          description: Not ready
  /v1/auth/session:
    get:
      summary: Validate user JWT and return session identity
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated session
        "401":
          description: Invalid token
  /v1/booking/confirm:
    post:
      summary: Confirm booking in Convex
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_type, date, time]
              properties:
                event_type:
                  type: string
                date:
                  type: string
                time:
                  type: string
      responses:
        "200":
          description: Booking confirmed
        "400":
          description: Validation failed
        "401":
          description: Unauthorized
        "500":
          description: Internal error
  /v1/payments/webhooks/stripe:
    post:
      summary: Receive Stripe webhook events
      parameters:
        - in: header
          name: x-oft-signature
          required: true
          schema:
            type: string
        - in: header
          name: x-oft-timestamp
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [event_id, event_type, status]
              properties:
                event_id:
                  type: string
                event_type:
                  type: string
                status:
                  type: string
                checkout_session_id:
                  type: string
                payment_intent_id:
                  type: string
      responses:
        "200":
          description: Processed or duplicated
        "401":
          description: Invalid signature
        "500":
          description: Internal error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT